springDb2/springtx/src/main/java/hello/springtx/propagation/MemberRepository.java
package hello.springtx.propagation;

import jakarta.persistence.EntityManager;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;

import java.util.Optional;

@Slf4j
@Repository
@RequiredArgsConstructor
public class MemberRepository {

    private final EntityManager em;

    @Transactional
    public void save(Member member){
        log.info("member 저장");
        em.persist(member);
    }

    public Optional<Member> find(String username){
        // pk 인 id 로 찾는 경우는 em.find() 로 찾으면되는데 여기서는 username으로 찾아 만들어서 직접 쿼리 만들어야함
       // jpql 을 사용해야함
        return em.createQuery("select m from Member m where m.username=:username",Member.class)
                .setParameter("username",username)
                .getResultList().stream().findAny();
    }
}
